Desktop/MR/requirements.txt
flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
pyodbc==5.0.1
